Move delegate methods to allow override
